Only trigger build in travis if there isn't one queued alread
The logic being that if a build is queued (behind a running build), this particular change
will be seen by the queued build, so we don't need yet another build after.

Travis Support confirmed ''auto-cancellation is only triggered when the event is a push or pull request. The assumption is that builds triggered via the API are definitely required, and therefore are not matched against the auto-cancellation condition.''